renames liteSiteCTA to liteSiteSummary and updates snapshots
