+++ src/app/components/LiteSiteSummary/index.test.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { LITE_ATI_CLICK_TRACKING } from '#app/lib/analyticsUtils/analytics.const';
 import { render } from '../react-testing-library-with-providers';
-import LiteSiteCta from '.';
+import LiteSiteSummary from '.';
 
-describe('LiteSiteCTA', () => {
+describe('LiteSiteSummary', () => {
   it('Should have a hidden strong element with lite site identifier.', () => {
-    const { container } = render(<LiteSiteCta />);
+    const { container } = render(<LiteSiteSummary />);
     const strongText = container.querySelector('strong');
     expect(strongText?.innerHTML).toBe('Data saving version');
     expect(strongText).toHaveAttribute('hidden');
   });
   it('Should have a CTA link to the main site.', () => {
-    const { container } = render(<LiteSiteCta />);
+    const { container } = render(<LiteSiteSummary />);
     const [ctaText] = container.querySelectorAll('a span');
     const [ctaLink] = container.querySelectorAll(
       'a[href="https://www.test.bbc.com/news/articles/c0g992jmmkko"]',
@@ -20,7 +20,7 @@ describe('LiteSiteCTA', () => {
     expect(ctaLink).toBeTruthy();
   });
   it('Should have a CTA link for more information.', () => {
-    const { container } = render(<LiteSiteCta />);
+    const { container } = render(<LiteSiteSummary />);
     const [, ctaText] = container.querySelectorAll('a span');
     const [ctaLink] = container.querySelectorAll(
       'a[href="https://www.test.bbc.com/news/articles/c0g992jmmkko"]',
@@ -33,11 +33,11 @@ describe('LiteSiteCTA', () => {
 
   it(`Should have click tracking on the 'Back to canonical' link.`, () => {
     process.env.SIMORGH_ATI_BASE_URL = 'https://logws1363.ati-host.net?';
-    const { container } = render(<LiteSiteCta />, { isLite: true });
+    const { container } = render(<LiteSiteSummary />, { isLite: true });
 
     const [ctaLink] = container.querySelectorAll('a');
     const atiUrl = ctaLink.getAttribute(LITE_ATI_CLICK_TRACKING);
 
-    expect(atiUrl).toContain('lite-site-cta');
+    expect(atiUrl).toContain('lite-site-summary');
   });
 });
