+++ src/app/legacy/containers/Header/index.jsx
@@ -9,7 +9,7 @@ import {
   TOPIC_PAGE,
   ERROR_PAGE,
 } from '#app/routes/utils/pageTypes';
-import LiteSiteCta from '#app/components/LiteSiteCta';
+import LiteSiteCta from '#app/components/LiteSiteSummary';
 import { ServiceContext } from '../../../contexts/ServiceContext';
 import ConsentBanner from '../ConsentBanner';
 import NavigationContainer from '../Navigation';
